move lazy promises to common
