+++ common/util.ts
@@ -866,3 +866,41 @@ export function toPropMap<T extends object, U extends PickProps<T, string>>(
 }
 
 type PickProps<T, U = string> = keyof { [key in keyof T as T[key] extends U ? key : never]: T[key] }
+
+export function lazyPromise<T = any>() {
+  const parts = {
+    resolve: (result: T) => {},
+    reject: (error: any) => {},
+    promise: {} as any as Promise<{ result?: T; error?: any }>,
+  }
+
+  parts.promise = new Promise<{ result?: T; error?: any }>((resolve, _reject) => {
+    parts.resolve = (result: any) => {
+      resolve({ result, error: undefined })
+    }
+    parts.reject = (error: any) => {
+      resolve({ result: undefined, error })
+    }
+  })
+
+  return parts
+}
+
+export function lazySimplePromise<T = any>() {
+  const parts = {
+    resolve: (result: T) => {},
+    reject: (error: any) => {},
+    promise: {} as any as Promise<T>,
+  }
+
+  parts.promise = new Promise<T>((resolve, _reject) => {
+    parts.resolve = (result: any) => {
+      resolve(result)
+    }
+    parts.reject = (error: any) => {
+      resolve(error)
+    }
+  })
+
+  return parts
+}
