+++ web/store/data/bot-generate.ts
@@ -26,7 +26,7 @@ import { replaceTags } from '/common/presets/templates'
 import { getServiceTempConfig } from '/web/shared/adapter'
 import { getActiveBots } from '/web/pages/Chat/util'
 import iconv from 'iconv-lite'
-import { genApi, lazyPromise } from './inference'
+import { genApi } from './inference'
 import { isDefaultPreset } from '/common/default-preset'
 import { ThirdPartyFormat } from '/common/adapters'
 import { localEmit } from '../socket'
@@ -37,6 +37,7 @@ import { getProvider } from '../preset-context'
 import { getLocalPayload, getStoppingStrings } from '/common/requests/payloads'
 import { sanitiseAndTrim } from '/common/requests/util'
 import { toastStore } from '../toasts'
+import { lazyPromise } from '/common/util'
 
 iconv.enableStreamingAPI(require('stream'))
